Add magical :prerender-keep attr for script tags
This commit introduces an experimental feature. Normally, the
prerendering code filters out script tags from the generated html,
because scripts created dynamically would be loaded twice if the page is
prerendered and those are included in the prerendered html.

For example:

    (page "index.html")

    (html
      (head
        (script "console.log('hello world')"))
      (body
        (h1 "check the console")))

If script tags were not filtered out of the prerendered HTML it would
look something like this:

    <!doctype html>
    <html>
      <head>
        <script>console.log('hello world')</script>
      </head>
      <body>
        <h1>check the console</h1>
        <script src="main.js"></script>
      </body>
    </html>

Notice the script tag inserted by the Hoplon compiler at the end of the
body, and the script tag created by the page Hoplon application code
when the page is loaded in PhantomJS. This is the contents of the HTML
file the browser will load.

So when this page is loaded in the client it will first run the script
in the <head>, render the <body>, and then load and run the main.js
script at the end of the <body>. The main.js script is the Hoplon
program, which will rebuild the DOM, recreating the prerendered content
including the script tag in the head. This is an issue because "hello
world" will be logged twice: once when the prerendered script tag is
run and then once more when Hoplon recreates that script tag at runtime.

The prerender task addresses this issue and solves it with a simple,
direct method: it just removes all script tags from the prerendered
HTML (except for the one that loads your main.js compiled Hoplon source,
of course).

But this creates other problems, namely third-party scripts that need
to run synchronously before the DOM is ready (perhaps they use
document.write(), perhaps they load other scripts, etc.). An example of
such a thing is Google Analytics (GA). We need a way to include these
types of scripts in the prerendered HTML, and exclude them when Hoplon
rebuilds the DOM at runtime.

We already added a feature that lets you detect at runtime whether the
page is being prerendered or not, via the #'hoplon.core/prerendering?
var. This commit adds a special, somewhat magical attribute you can use
with script tags that will tell the prerender task not to filter them
out, the :prerender-keep attribute.

Consider GA in a simple page, for example:

    (page "index.html")

    (html
      (head
        (when prerendering?
          (script :prerender-keep "keep"
            "(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject...")))
      (body
        (h1 "hello world")))

This will include the GA script only in the prerendered HTML, and will
not recreate it when Hoplon builds the page. You should be able to see
the GA script in the HTML source (eg. via "view page source" in the
Chrome context menu), and you *shouldn't* be able to see it in the
generated DOM (eg. via the "elements" tab in Chrome dev tools).